Add mocking interface
