+++ src/cypress-helper.spec.cy.ts
@@ -1,4 +1,4 @@
-import { CypressHelper, loggable, match } from ".";
+import { CypressHelper, loggable } from ".";
 import { then } from "./assertable";
 
 describe("cypress helper tests", () => {
@@ -302,176 +302,6 @@ describe("cypress helper tests", () => {
 
   it("should get enabled element status", () => {
     then(get.elementByTestId("submit")).shouldBeEnabled();
-  });
-
-  describe("stubs and spies", () => {
-    it("should spy on function", () => {
-      const obj = {
-        func: (param: number) => {}
-      };
-      given.spyOnObject(obj, "func");
-      obj.func(3);
-      then(get.spyFromFunction(obj.func)).shouldHaveBeenCalledWith(3);
-    });
-
-    it("should partially match spy params", () => {
-      const obj = {
-        func: (param: Object) => {}
-      };
-      given.spyOnObject(obj, "func");
-      obj.func({ shelly: "go", inner: { attr: "value" } });
-      then(get.spyFromFunction(obj.func)).shouldHaveBeenCalledWithMatch(
-        match({ inner: { attr: "value" } })
-      );
-    });
-
-    it("should stub function", () => {
-      let func = () => 5;
-      func = given.stub().returns(7);
-      expect(func()).to.eq(7);
-    });
-
-    it("should stub function with alias", () => {
-      let func = (input: number) => input;
-      func = given.stub("func");
-      func(7);
-      then(get.stub("func")).shouldHaveBeenCalledWith(7);
-    });
-
-    it("should stub object function", () => {
-      const obj = {
-        func: () => 5
-      };
-      given.stubObjectMethod(obj, "func").returns(7);
-      then(get.assertableStub(obj.func())).shouldEqual(7);
-    });
-
-    it("should stub object getter", () => {
-      const obj = {
-        get count() {
-          return 5;
-        }
-      };
-      given.stubObjectMethod(obj, "count").get(() => 7);
-      expect(obj.count).to.eq(7);
-    });
-
-    describe("stubbing class", () => {
-      class MyClass {
-        constructor(input: number) {}
-        func(input: number, text: string) {
-          console.log(text);
-          return input;
-        }
-        property: number = 3;
-        get getter(): number {
-          return this.property;
-        }
-        set setter(value: number) {
-          this.property = value;
-        }
-
-        propertyFunc = (int: number) => int;
-
-        asyncPropertyFunc = async (int: number) => await Promise.resolve(int);
-        async asynFunc() {
-          return await Promise.resolve(9);
-        }
-      }
-
-      it("should assert stub async function calls", async () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.asynFunc.returns(Promise.resolve(7));
-        mockMyClass.propertyFunc(3);
-        then(get.assertableStub(mockMyClass.asynFunc)).shouldHaveBeenCalledWith(
-          3
-        );
-      });
-
-      it("should assert stub async function calls using overrides", async () => {
-        const stub = given.stub().returns(Promise.resolve(7));
-        const mockMyClass = given.stubbedInstance(MyClass, { asynFunc: stub });
-        mockMyClass.propertyFunc(3);
-        then(get.assertableStub(mockMyClass.asynFunc)).shouldHaveBeenCalledWith(
-          3
-        );
-      });
-
-      it("should stub async function", async () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.asynFunc.returns(Promise.resolve(7));
-        mockMyClass.propertyFunc(3);
-        expect(await mockMyClass.asynFunc()).to.eq(7);
-      });
-
-      it("should stub async function using overrides", async () => {
-        const stub = given.stub().returns(Promise.resolve(7));
-        const mockMyClass = given.stubbedInstance(MyClass, { asynFunc: stub });
-        mockMyClass.propertyFunc(3);
-        expect(await mockMyClass.asynFunc()).to.eq(7);
-      });
-
-      it("should stub async property function", async () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        await mockMyClass.asyncPropertyFunc(3);
-        then(
-          get.assertableStub(mockMyClass.asyncPropertyFunc)
-        ).shouldHaveBeenCalledWith(3);
-      });
-
-      it("should override class property Function", () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.propertyFunc.returns(7);
-        expect(mockMyClass.propertyFunc(3)).to.eq(7);
-      });
-
-      it("should stub class property function", () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.propertyFunc(3);
-        then(
-          get.assertableStub(mockMyClass.propertyFunc)
-        ).shouldHaveBeenCalledWith(3);
-      });
-
-      it("should override class property", () => {
-        const mockMyClass = given.stubbedInstance(MyClass, { property: 5 });
-        expect(mockMyClass.property).to.eq(5);
-      });
-
-      it("should override class getter", () => {
-        const mockMyClass = given.stubbedInstance(MyClass, { getter: 5 });
-        expect(mockMyClass.getter).to.eq(5);
-      });
-
-      it("should stub class setter with override", () => {
-        const mockMyClass = given.stubbedInstance(MyClass, { property: 8 });
-        mockMyClass.setter = 5;
-        expect(mockMyClass.property).to.eq(8);
-      });
-
-      it("should assert stub class setter calls", () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.setter = 5;
-        then(get.assertableStub(mockMyClass.setter)).shouldHaveBeenCalledWith(
-          5
-        );
-      });
-
-      it("should stub class", () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.func(5, "whatever");
-        then(get.assertableStub(mockMyClass.func)).shouldHaveBeenCalledWith(
-          5,
-          "whatever"
-        );
-      });
-
-      it("should stub class function return value", () => {
-        const mockMyClass = given.stubbedInstance(MyClass);
-        mockMyClass.func.returns(7);
-        expect(mockMyClass.func(5, "whatever")).to.eq(7);
-      });
-    });
   });
 
   it("should get element by text", () => {
